refactor(api): :recycle: Separate list item api routes from client routes
Refs #12
