+++ routes/lists/[id]/delete.ts
@@ -1,6 +1,6 @@
 import List from "lists/mod.ts";
 import ListItem from "list-items/mod.ts";
-import ListItemStorage from "../../list-items/db.ts";
+import ListItemStorage from "../../api/list-items/db.ts";
 import Storage from "../db.ts";
 import { Handlers } from "$fresh/server.ts";
